+++ frontend/src/app/(appLayout)/search/users/page.tsx
@@ -5,8 +5,8 @@ import PageContainer from "@/components/common/containers/PageContainer";
 import PageTitle from "@/components/common/ui/PageTitle";
 import type { User } from "@/types/User";
 import { baseUrl } from "@/constants";
-import ButtonAddFriend from "@/components/common/ui/ButtonAddFriend";
-import UserAvatar from "@/components/common/ui/UserAvatar";
+import ButtonAddFriend from "@/components/common/ui/buttons/ButtonAddFriend";
+import UserAvatar from "@/components/common/ui/image/UserAvatar";
 import { MapPinHouse, FolderGit } from "lucide-react";
 import { send } from "process";
 import { auth } from "@/auth";
