+++ frontend/src/components/common/layout/Navbar.tsx
@@ -14,13 +14,13 @@ import { useWindowWidth } from "@/hooks/useWindowWidth";
 import { useSidebarsContext } from "@/context/SidebarsContext";
 
 //Components
-import ToggleTheme from "../ui/ToggleTheme";
+import ToggleTheme from "../ui/buttons/ToggleTheme";
 //UI
 import { FaGithub, FaLinkedin } from "react-icons/fa";
 import { Button } from "@/components/ui/button";
 import logo from "/public/logo.png";
 import { contributors } from "@/constants";
-import UserAvatar from "@/components/common/ui/UserAvatar";
+import UserAvatar from "@/components/common/ui/image/UserAvatar";
 import NavbarLinks from "./NavbarLinks";
 import { Menu, SquareUserRound } from "lucide-react";
 import NavbarAsozialMenu from "./NavbarAsozialMenu";
