+++ frontend/src/components/common/layout/NavbarAccountMenu.tsx
@@ -8,11 +8,11 @@ import {
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
 import Link from "next/link";
-import UserAvatar from "../ui/UserAvatar";
+import UserAvatar from "../ui/image/UserAvatar";
 import { User, Settings, LogOut } from "lucide-react";
 import { signOut } from "@/actions";
 import { Button } from "@/components/ui/button";
-import ToggleTheme from "../ui/ToggleTheme";
+import ToggleTheme from "../ui/buttons/ToggleTheme";
 
 type NavbarAccountMenuProps = {
   userId: string;
