+++ frontend/src/components/common/sidebar/SidebarFriendsList.tsx
