+++ frontend/src/components/common/sidebar/SidebarTitle.tsx
