+++ frontend/src/components/common/ui/ImageUploader.tsx
@@ -6,9 +6,9 @@ import { uploadFile } from "@/actions";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { cn } from "@/lib/utils";
-import LoadingSpinner from "./LoadingSpinner";
+import LoadingSpinner from "./loading/LoadingSpinner";
 import { SquareCheckBig, Upload } from "lucide-react";
-import LoadingTextButton from "./LoadingTextButton";
+import LoadingTextButton from "./loading/LoadingTextButton";
 import ErrorMessage from "./ErrorMessage";
 
 export type ImageT = {
