+++ frontend/src/components/common/ui/icons/CancelIcon.tsx
