+++ frontend/src/components/common/ui/icons/DashboardIcon.tsx
