+++ frontend/src/components/common/ui/icons/DeleteIcon.tsx
