+++ frontend/src/components/common/ui/icons/ProjectCreateIcon.tsx
