+++ frontend/src/components/common/ui/icons/ProjectIcon.tsx
