+++ frontend/src/components/common/ui/icons/SearchIcon.tsx
