+++ frontend/src/components/common/ui/icons/ThumbsDownIcon.tsx
