+++ frontend/src/components/common/ui/icons/ThumbsUpIcon.tsx
