+++ frontend/src/components/common/ui/image/UserAvatar.tsx
