+++ frontend/src/components/dashboard/DashboardContainer.tsx
@@ -4,7 +4,7 @@ import { auth } from "@/auth";
 import FileUploader from "../common/ui/ImageUploader";
 import { getUserFriendStatuses } from "@/actions/friendships.server/getUserFriendStatuses";
 import { get } from "http";
-import UserAvatar from "../common/ui/UserAvatar";
+import UserAvatar from "../common/ui/image/UserAvatar";
 import { Button } from "../ui/button";
 import { UserRoundCheck, UserRoundX } from "lucide-react";
