+++ frontend/src/components/project/SyncRepoButton.tsx
@@ -1,6 +1,6 @@
 import { useFormStatus } from "react-dom";
 import { Button } from "../ui/button";
-import LoadingTextButton from "../common/ui/LoadingTextButton";
+import LoadingTextButton from "../common/ui/loading/LoadingTextButton";
 
 function SyncRepoButton() {
   const { pending } = useFormStatus();
