+++ frontend/src/components/project/project-form/Title.tsx
@@ -14,7 +14,7 @@ import { SquareCheck } from "lucide-react";
 //Lib
 import { Control, FieldErrors, UseFormSetValue } from "react-hook-form";
 import SuccessMessage from "@/components/common/ui/SuccessMessage";
-import LoadingTextButton from "@/components/common/ui/LoadingTextButton";
+import LoadingTextButton from "@/components/common/ui/loading/LoadingTextButton";
 
 type TitleProps = {
   errors: FieldErrors<Inputs>;
