+++ frontend/src/components/user/UserSidebar.tsx
@@ -5,12 +5,12 @@ import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import { useSession } from "next-auth/react";
 import { Button } from "../ui/button";
 import { signOut } from "@/actions";
-import ButtonBack from "../common/ui/ButtonBack";
-import SidebarTitle from "../common/ui/SidebarTitle";
-import SidebarFriendsList from "./SidebarFriendsList";
+import ButtonBack from "../common/ui/buttons/ButtonBack";
+import SidebarTitle from "../common/sidebar/SidebarTitle";
+import SidebarFriendsList from "../common/sidebar/SidebarFriendsList";
 import SidebarButtons from "../common/sidebar/SidebarButtons";
 import { SearchIcon } from "lucide-react";
-import MatchIcon from "../common/ui/MatchIcon";
+import MatchIcon from "../common/ui/icons/MatchIcon";
 
 function UserSidebar() {
   const { userSidebarRef, isUserSidebarOpen, projectSidebarRef } =
